+++ isik/django/apps/common/db/__init__.py
@@ -1,12 +1,14 @@
 from isik.django.apps.common.db import lookups  # noqa: F401
-from isik.django.apps.common.db.history import track_events
+from isik.django.apps.common.db.history import event_model_for, history_middleware_installed, track_events
 from isik.django.apps.common.db.models import BaseModel
 from isik.django.apps.common.db.orm import get_object_or_none, starts_with
 
 
 __all__ = [
     "BaseModel",
+    "event_model_for",
     "get_object_or_none",
+    "history_middleware_installed",
     "starts_with",
     "track_events",
 ]