+++ isik/django/drf/serializers/__init__.py
@@ -6,6 +6,7 @@ from isik.django.drf.serializers.conditional_serializer import (
 )
 from isik.django.drf.serializers.create_only import CreateOnlyFieldsMixin
 from isik.django.drf.serializers.flattened_one_to_one import FlattenedOneToOneMixin
+from isik.django.drf.serializers.history import generic_history_serializer
 from isik.django.drf.serializers.meta_combining import MetaCombiningMixin
 from isik.django.drf.serializers.registry import ModelSerializerRegistryMixin
 from isik.django.drf.serializers.request_context import RequestContextMixin
@@ -21,6 +22,7 @@ __all__ = [
     "ModelSerializerRegistryMixin",
     "RequestContextMixin",
     "WriteOnlyFieldsMixin",
+    "generic_history_serializer",
     "relational_serializer",
     "serializer_method_include",
 ]