+++ isik/django/drf/viewsets/__init__.py
@@ -1,6 +1,7 @@
 from isik.django.drf.viewsets.action_serializer_class import ActionSerializerClassMixin
 from isik.django.drf.viewsets.base import BaseModelViewSet
 from isik.django.drf.viewsets.filterset import FilterSetMixin
+from isik.django.drf.viewsets.history import HistoryMixin, context_filter
 from isik.django.drf.viewsets.ordering import ReverseOrderingMixin
 from isik.django.drf.viewsets.protected_destroy import ProtectedDestroyMixin
 from isik.django.drf.viewsets.registry import ViewSetRegistryMixin
@@ -11,8 +12,10 @@ __all__ = [
     "ActionSerializerClassMixin",
     "BaseModelViewSet",
     "FilterSetMixin",
+    "HistoryMixin",
     "ProtectedDestroyMixin",
     "ReverseOrderingMixin",
     "ViewSetRegistryMixin",
+    "context_filter",
     "none_during_schema_generation",
 ]